enh(#50): Path following force
- Also re-run clang-tidy and clean most warnings, including one about integer
  division in abort_probability which probably caused less task aborts than
  there should have been otherwise.

- Cannot test yet, because the necessary controller in SILICON is not
  implemented, but should be fairly close to correct.

+semver: enh

+++ include/cosm/metrics/collector_registerer.hpp
@@ -319,9 +319,9 @@ class collector_registerer_impl<std::tuple<Ts...>> : public rer::client<collecto
     auto append_it = mc_config->append.enabled.find(xml_name);
     auto truncate_it = mc_config->truncate.enabled.find(xml_name);
     auto create_it = mc_config->create.enabled.find(xml_name);
-    uint sum = (append_it != mc_config->append.enabled.end()) +
-               (truncate_it != mc_config->truncate.enabled.end()) +
-               (create_it != mc_config->create.enabled.end());
+    uint sum = static_cast<uint>(append_it != mc_config->append.enabled.end()) +
+               static_cast<uint>(truncate_it != mc_config->truncate.enabled.end()) +
+               static_cast<uint>(create_it != mc_config->create.enabled.end());
     ER_ASSERT(
         sum <= 1,
         "Collector '%s' present in more than 1 collector group in XML file",
