+++ src/metrics/blocks/transport_metrics_collector.cpp
@@ -105,12 +105,12 @@ boost::optional<std::string> transport_metrics_collector::csv_line_build(void) {
 void transport_metrics_collector::collect(const rmetrics::base_metrics& metrics) {
   auto& m = dynamic_cast<const transport_metrics&>(metrics);
   ++m_interval.transported;
-  m_interval.cube_transported += (repr::block_type::ekCUBE == m.type());
-  m_interval.ramp_transported += (repr::block_type::ekRAMP == m.type());
+  m_interval.cube_transported += static_cast<uint>(repr::block_type::ekCUBE == m.type());
+  m_interval.ramp_transported += static_cast<uint>(repr::block_type::ekRAMP == m.type());
 
   ++m_cum.transported;
-  m_cum.cube_transported += (repr::block_type::ekCUBE == m.type());
-  m_cum.ramp_transported += (repr::block_type::ekRAMP == m.type());
+  m_cum.cube_transported += static_cast<uint>(repr::block_type::ekCUBE == m.type());
+  m_cum.ramp_transported += static_cast<uint>(repr::block_type::ekRAMP == m.type());
 
   m_interval.transporters += m.total_transporters();
   m_cum.transporters += m.total_transporters();
